+++ ComputerGraphics/MeshLoader.h
@@ -10,7 +10,7 @@ private:
 	CustomTriangle* m_triangles;
 public:
 	MeshLoader(std::string filepath);
-	void load(int scale, float xOffset, float yOffset);
+	void load(float scale, float xOffset, float yOffset);
 	const CustomTriangle* getTriangles();
 	int getTriangleCount();
 };
